dateWithAi_python/requirements.txt
openai==1.35.13
psycopg2-binary==2.9.7
pgvector==0.2.5
python-dotenv==1.0.0
numpy==1.24.3
tiktoken==0.7.0
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
httpx==0.25.0